+++ resources/docker/app/src/public/assets/js/modules/main.js
@@ -117,17 +117,17 @@ let es6exports = {};
                     this.firstChild.textContent = "analyze ";
                     toggleClass(this.firstElementChild,"fa-expand",true)
                         .toggleClass("fa-compress",false);
-                    toggleClass(panelWrapper,"smaller",false);
+                    /* toggleClass(panelWrapper,"smaller",false); // uncomment for enabling shrinking*/
                     toggleClass(side,"expanded",false);
                 } else {
                     this._status = true;
                     this.firstChild.textContent = "hide ";
                     toggleClass(this.firstElementChild,"fa-expand",false)
                         .toggleClass("fa-compress",true);
-                    toggleClass(panelWrapper,"smaller",true);
+                    /* toggleClass(panelWrapper,"smaller",true); // uncomment for enabling shrinking*/
                     toggleClass(side,"expanded",true);
                 }
-                zoomPanel();
+               /*  zoomPanel(); //uncomment for enabling shrinking*/
             },false);
             uploadButton.addEventListener("click",function(e){
                 const rgxNexus = /^https:\/\/dl\.dnanex\.us\//gi;
@@ -184,55 +184,6 @@ let es6exports = {};
                 selectedPanel = panels[index - 1];
                 zoomPanel(); */
             },false);
-            actionButton.addEventListener("click", function(){
-                if(this._disabled){return}
-                this._disabled = true;
-                if (!this._igvLoaded){
-                    const that = this;
-                    Swal.fire({
-                        title: "Loading IGV",
-                        allowEscapeKey: false,
-                        allowOutsideClick: false,
-                        backdrop: false,
-                        didOpen: () => {
-                            Swal.showLoading();
-                        }
-                    });
-                    taskq.load("/static/js/igv.js").then(function(res){
-                        that._disabled = false;
-                        that._igvLoaded = true;
-                        Swal.close();
-                        that.click();
-                    });
-                    return;
-                }
-                let oIGV = createIGVObject();
-                if (oIGV instanceof Error) {
-                    Swal.fire("make sure you select a bam file");
-                    this._disabled = false;
-                    return;
-                }
-                Promise.resolve()
-                .then(() => {
-                    if(igv.browser) {
-                        return igv.browser.loadSessionObject(oIGV)
-                        .then(() => {
-                            console.log("igv redrawn");
-                            this._disabled = false;
-                        });
-                    }
-                    return igv.createBrowser(selectedPanel.firstElementChild, oIGV)
-                    .then(browser => {
-                        igv.browser = browser;
-                        console.log("igv rendered");
-                        this._disabled = false;
-                    })
-                })
-                .catch(err => {
-                    console.log(err);
-                    this._disabled = false;
-                });
-            }, false);
             deleteButtons.forEach((slave, master) => {
                 slave.addEventListener("click", async function(){
                     if(!master.value){
@@ -260,6 +211,18 @@ let es6exports = {};
             });
         })
         .then(function(){
+            const doneObj = {value: void(0), done: false};
+            Promise.all([
+                import("./main-annotation-indexing.js").then(f => ({enableAnnotations: f.default}))
+            ])
+            .then(objs => Object.assign({}, ...objs))
+            .then(dyImports => {
+                doneObj.value = dyImports;
+                doneObj.done = true;
+            })
+            return doneObj;
+        })
+        .then(function(dyImports){
             const evtSource 
                 = es6exports.evtSource 
                 = new EventSource('/estream/subscribe');
@@ -311,6 +274,7 @@ let es6exports = {};
                     headerTemplate.textContent = "File size: " + bytesToHuman(oPayload[this.value]);
                 });
             });
+            dyImports.enableAnnotations(evtSource);
             evtSource.addEventListener("error", function(e){
                 if (connectionLost) {return}
                 connectionLost = true;
@@ -421,7 +385,10 @@ let es6exports = {};
              .export(themeButton, "themeButton")
              .export(helpButton, "helpButton")
              .export(toolsButton, "toolsButton")
-             .export(firstHexGrid, "firstHexGrid");
+             .export(firstHexGrid, "firstHexGrid")
+             .export(actionButton, "actionButton")
+             .export(selectedPanel, "selectedPanel") //panels[0] is for IGV by default
+             .export(dropdown, "dropdown"); //orignally used for bam files
         taskq._exportPersist.helpButton = helpButton;
         taskq._exportPersist.expandButton = expandButton;
         taskq._exportPersist.toolsButton = toolsButton;